requirements.txt
flask==0.10.1
gunicorn==19.3.0
pyyaml==3.11
python-slugify==1.1.3
tinydb==3.1.1
ujson==1.35
